+++ ibis/backends/bigquery/__init__.py
@@ -7,12 +7,13 @@ import google.cloud.bigquery  # noqa: F401, fail early if bigquery is missing
 import pydata_google_auth
 
 import ibis.common.exceptions as com
-from ibis.bigquery.client import BigQueryClient
-from ibis.bigquery.compiler import dialect
 from ibis.config import options  # noqa: F401
 
+from .client import BigQueryClient
+from .compiler import dialect
+
 try:
-    from ibis.bigquery.udf import udf  # noqa: F401
+    from .udf import udf
 except ImportError:
     pass
 
@@ -32,7 +33,7 @@ def compile(expr, params=None):
     ibis.expr.types.Expr.compile
 
     """
-    from ibis.bigquery.compiler import to_sql
+    from .compiler import to_sql
 
     return to_sql(expr, dialect.make_context(params=params))
