+++ ibis/backends/bigquery/compiler.py
@@ -13,7 +13,6 @@ import ibis.expr.lineage as lin
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
 import ibis.sql.compiler as comp
-from ibis.bigquery.datatypes import ibis_type_to_bigquery_type
 from ibis.impala import compiler as impala_compiler
 from ibis.impala.compiler import (
     ImpalaSelect,
@@ -22,6 +21,8 @@ from ibis.impala.compiler import (
     fixed_arity,
     unary,
 )
+
+from .datatypes import ibis_type_to_bigquery_type
 
 
 class BigQueryUDFNode(ops.ValueOp):
