+++ ibis/backends/bigquery/tests/test_datatypes.py
@@ -3,7 +3,8 @@ from multipledispatch.conflict import ambiguities
 from pytest import param
 
 import ibis.expr.datatypes as dt
-from ibis.bigquery.datatypes import (
+
+from .datatypes import (
     TypeTranslationContext,
     UDFContext,
     ibis_type_to_bigquery_type,
