+++ ibis/backends/bigquery/tests/udf/test_udf_execute.py
@@ -7,8 +7,9 @@ from pytest import param
 
 import ibis
 import ibis.expr.datatypes as dt
-from ibis.bigquery import udf  # noqa: E402
 from ibis.compat import PY38
+
+from . import udf  # noqa: E402
 
 pytest.importorskip('google.cloud.bigquery')
 
@@ -26,7 +27,7 @@ DATASET_ID = 'testing'
 
 @pytest.fixture(scope='module')
 def client():
-    from ibis.bigquery.tests.conftest import connect
+    from .tests.conftest import connect
 
     return connect(PROJECT_ID, DATASET_ID)
