+++ ibis/backends/bigquery/udf/__init__.py
@@ -6,11 +6,12 @@ import itertools
 import ibis.expr.datatypes as dt
 import ibis.expr.rules as rlz
 import ibis.udf.validate as v
-from ibis.bigquery.compiler import BigQueryUDFNode, compiles
-from ibis.bigquery.datatypes import UDFContext, ibis_type_to_bigquery_type
-from ibis.bigquery.udf.core import PythonToJavaScriptTranslator
 from ibis.compat import PY38  # noqa: F401
 from ibis.expr.signature import Argument as Arg
+
+from ..compiler import BigQueryUDFNode, compiles
+from ..datatypes import UDFContext, ibis_type_to_bigquery_type
+from ..udf.core import PythonToJavaScriptTranslator
 
 __all__ = ('udf',)
