--- ibis/bigquery/udf/__init__.py
@@ -1,1 +1,0 @@
-from ibis.bigquery.udf.api import udf  # noqa: F401
